python-app/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.25
psycopg2-binary==2.9.1
python-dotenv==0.19.1
requests==2.26.0  
beautifulsoup4==4.9.3  
selenium==3.141.0